+++ src/components/Typography/Header.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { StyleSheet } from 'react-native'
 
-import Fonts from '../themes/fonts'
+import Fonts from '../../themes/fonts'
 
-import Text, { TextProps } from './Text'
+import { Text, TextProps } from './Text'
 
 export type HeaderProps = TextProps
 
@@ -14,9 +14,10 @@ const styles = StyleSheet.create( {
   },
 } )
 
-const Header = ( {
+/**
+ * Styled header text component.
+ */
+export const Header = ( {
   style,
   ...props
 }: HeaderProps ) => <Text style={[ styles.header, style ]} {...props} />
-
-export default Header
