+++ src/components/Typography/Text.tsx
@@ -1,8 +1,8 @@
 import React, { ReactNode } from 'react'
 import { StyleSheet, Text as NativeText, TextProps as NativeTextProps } from 'react-native'
 
-import Colours from '../themes/colours'
-import Fonts from '../themes/fonts'
+import Colours from '../../themes/colours'
+import Fonts from '../../themes/fonts'
 
 export type TextProps = NativeTextProps & { children: ReactNode }
 
@@ -13,9 +13,10 @@ const styles = StyleSheet.create( {
   },
 } )
 
-const Text = ( {
+/**
+ * Themed text component.
+ */
+export const Text = ( {
   style,
   ...props
 }: TextProps ) => <NativeText style={[ styles.text, style ]} {...props} />
-
-export default Text
