+++ src/components/Typography/index.ts
@@ -1,0 +1,3 @@
+export * from './Header'
+
+export * from './Text'
